+++ src/app/products/components/product-list/product-list.component.ts
@@ -4,6 +4,8 @@ import { ProductItem } from '../../models/product.model';
 import { ProductsService } from '../../services/products.service';
 import { CartService } from '../../../cart/services/cart.service';
 
+import { Observable, of, Subscription } from 'rxjs';
+
 @Component({
   selector: 'app-product-list',
   templateUrl: './product-list.component.html',
@@ -12,19 +14,55 @@ import { CartService } from '../../../cart/services/cart.service';
 export class ProductListComponent implements OnInit {
 
   products: Array<ProductItem> = [];
+  productsObservable: Observable<ProductItem[]> = of([]);
+
+  productsPromise: Promise<ProductItem[]> = new Promise(()=>[]);
+  productsFromPromise : Array<ProductItem> = [];
 
   constructor(private productService: ProductsService,
               private cartService: CartService) {
-
+  
   }
 
   ngOnInit(): void {
     this.products = this.productService.getProducts();
+
+    this.productsObservable = this.productService.getProductsAsyncObservable();
+
+    this.productsPromise = this.productService.getProductsPromise();
+    
+    // this.productsPromise.then(data=>{
+    //   console.log('ngOnInit(): data: ' + data);
+    //   console.log('ngOnInit(): -product: ' + data[0].name);
+    //   this.productsFromPromise=data;
+    //   console.log('ngOnInit(): this.productsFromPromise: ' + this.productsFromPromise);
+    // });
   }
 
   onAddInCart(product: ProductItem): void {
 
     this.cartService.addProduct(product);
+  }
+
+  getProductsFromPromise(): ProductItem[] {
+    
+    // (async()=>{
+      
+    //   this.productsFromPromise = await this.productService.getProductsPromise();
+      
+    // });
+
+    // return this.productsFromPromise
+
+    // this.productsPromise = this.productService.getProductsPromise();
+    // this.productsPromise.then(data=>{
+    //   console.log('data: ' + data);
+    //   console.log('-product: ' + data[0].name);
+    //   this.productsFromPromise=data
+    // });
+
+    console.log('getProductsFromPromise(): this.productsFromPromise: ' + this.productsFromPromise);
+    return this.productsFromPromise;
   }
 
 }