+++ src/app/products/components/product/product.component.ts
@@ -1,6 +1,7 @@
 import { Component, EventEmitter, Input, Output, OnInit } from '@angular/core';
 
 import { ProductItem } from '../../models/product.model';
+import { Observable, Subscriber, of } from 'rxjs';
 
 @Component({
   selector: 'app-product',
@@ -10,15 +11,26 @@ import { ProductItem } from '../../models/product.model';
 export class ProductComponent implements OnInit {
 
   @Input()
-  products: Array<ProductItem> = [];
+  products: Array<ProductItem> | null = [];
+
+  // @Input()
+  // productsObservable: Observable<Array<ProductItem>> | undefined;
 
   @Output()
   productToBuy: EventEmitter<ProductItem> = new EventEmitter<ProductItem>();
+
+  productsForTable: Array<ProductItem> = [];
 
   headers: string[] = ['id', 'name', 'price', 'actions'];
 
   ngOnInit(): void {
 
+    console.log('this.products: ' + this.products);
+
+    if (this.products != null) {
+      this.productsForTable = this.products;
+    }
+        
   }
 
   onBuy(): void {